Remove allocation when getting 'status' label value
The vast majority of requests will have a 'valid' status code and so we
can use the http crate's `StatusCode` struct to represent them, which
provides a non-allocating `as_str()` method to get the status code label
used for built-in metrics.

In other cases, fall back to converting the status code to a string,
with an allocation.