Use an internal Registry for Rocket metrics
This registry is created by `PrometheusMetrics::with_registry` and is
private to each `PrometheusMetrics` instance, allowing multiple
`PrometheusMetrics` instances to share the same `extra_registry`.

Previously the fairing tried to register the internal metrics on the `extra_registry`,
which caused conflicts if the same registry was passed twice. This is now avoided
by using an internal registry for those metrics.

Both registries are gathered and serialized to the same buffer when the
/metrics endpoint is called.

Closes #14.